requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-Migrate==4.0.4
Flask-Limiter==3.3.1
python-dotenv==1.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20